Move package from `com` to `dev` and fix docs
