--- MaterialDialogLibrary/src/main/java/com/shreyaspatil/MaterialDialog/interfaces/DialogInterface.java
@@ -1,6 +1,0 @@
-package com.shreyaspatil.MaterialDialog.interfaces;
-
-public interface DialogInterface {
-    void cancel();
-    void dismiss();
-}
