--- MaterialDialogLibrary/src/main/java/com/shreyaspatil/MaterialDialog/interfaces/OnCancelListener.java
@@ -1,5 +1,0 @@
-package com.shreyaspatil.MaterialDialog.interfaces;
-
-public interface OnCancelListener {
-    void onCancel(DialogInterface dialogInterface);
-}
