--- MaterialDialogLibrary/src/main/java/com/shreyaspatil/MaterialDialog/interfaces/OnDismissListener.java
@@ -1,5 +1,0 @@
-package com.shreyaspatil.MaterialDialog.interfaces;
-
-public interface OnDismissListener {
-    void onDismiss(DialogInterface dialogInterface);
-}
