--- MaterialDialogLibrary/src/main/java/com/shreyaspatil/MaterialDialog/interfaces/OnShowListener.java
@@ -1,5 +1,0 @@
-package com.shreyaspatil.MaterialDialog.interfaces;
-
-public interface OnShowListener {
-    void onShow(DialogInterface dialogInterface);
-}
