+++ MaterialDialogLibrary/src/main/java/dev/shreyaspatil/MaterialDialog/interfaces/DialogInterface.java
@@ -1,0 +1,6 @@
+package dev.shreyaspatil.MaterialDialog.interfaces;
+
+public interface DialogInterface {
+    void cancel();
+    void dismiss();
+}
