+++ MaterialDialogLibrary/src/main/java/dev/shreyaspatil/MaterialDialog/interfaces/OnCancelListener.java
@@ -1,0 +1,5 @@
+package dev.shreyaspatil.MaterialDialog.interfaces;
+
+public interface OnCancelListener {
+    void onCancel(DialogInterface dialogInterface);
+}
