+++ MaterialDialogLibrary/src/main/java/dev/shreyaspatil/MaterialDialog/interfaces/OnDismissListener.java
@@ -1,0 +1,5 @@
+package dev.shreyaspatil.MaterialDialog.interfaces;
+
+public interface OnDismissListener {
+    void onDismiss(DialogInterface dialogInterface);
+}
