+++ MaterialDialogLibrary/src/main/java/dev/shreyaspatil/MaterialDialog/interfaces/OnShowListener.java
@@ -1,0 +1,5 @@
+package dev.shreyaspatil.MaterialDialog.interfaces;
+
+public interface OnShowListener {
+    void onShow(DialogInterface dialogInterface);
+}
