+++ app/src/androidTest/java/dev/shreyaspatil/MaterialDialogExample/ExampleInstrumentedTest.java
@@ -1,0 +1,27 @@
+package dev.shreyaspatil.MaterialDialogExample;
+
+import android.content.Context;
+
+import androidx.test.InstrumentationRegistry;
+import androidx.test.runner.AndroidJUnit4;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+
+import static org.junit.Assert.assertEquals;
+
+/**
+ * Instrumented test, which will execute on an Android device.
+ *
+ * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
+ */
+@RunWith(AndroidJUnit4.class)
+public class ExampleInstrumentedTest {
+    @Test
+    public void useAppContext() {
+        // Context of the app under test.
+        Context appContext = InstrumentationRegistry.getTargetContext();
+
+        assertEquals("com.shreyaspatil.MaterialDialog", appContext.getPackageName());
+    }
+}
