+++ app/src/main/res/layout/activity_main.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context="dev.shreyaspatil.MaterialDialogExample.MainActivity">
 
     <LinearLayout
         android:padding="16dp"
